requirements.txt
flask
python-dotenv
openai
twilio
uuid
